requirements.txt
langchain
beautifulsoup4
requests
flask
gunicorn
langchain-groq
python-dotenv
langchain-google-community
flask-cors